Revert "nothing"
This reverts commit 3fb6b3000b2cffe738c6db296f7be5fe2f211949.

+++ core/src/com/bitfire/postprocessing/effects/LensFlare2.java
@@ -217,7 +217,7 @@ public final class LensFlare2 extends PostProcessorEffect {
 	}
 
 	public int getGhosts () {
-		return lens.getGhosts();
+		return (int)lens.getGhosts();
 	}
 
 	public boolean isBlendingEnabled () {
